// src/types.ts
export enum ModelLangToLangId {
	bat = 'bat',
	cmd = 'bat',
	btm = 'bat',
	c = 'c',
	cs = 'csharp',
	cpp = 'cpp',
	cc = 'cpp',
	coffee = 'coffeescript',
	litcoffee = 'coffeescript',
	css = 'css',
	erl = 'erlang',
	hrl = 'erlang',
	go = 'go',
	hs = 'haskell',
	lhs = 'haskell',
	html = 'html',
	java = 'java',
	js = 'javascript',
	es6 = 'javascript',
	ipynb = 'jupyter',
	lua = 'lua',
	md = 'markdown',
	matlab = 'matlab',
	m = 'objective-c',
	mm = 'objective-c',
	pl = 'perl',
	pm = 'perl',
	php = 'php',
	ps1 = 'powershell',
	py = 'python',
	r = 'r',
	rdata = 'r',
	rds = 'r',
	rda = 'r',
	rb = 'ruby',
	rs = 'rust',
	scala = 'scala',
	sh = 'shellscript',
	sql = 'sql',
	swift = 'swift',
	tex = 'tex',
	ts = 'typescript',
	tsx = 'typescriptreact'
};

export interface ModelResult {
	languageId: ModelLangToLangId;
	confidence: number;
}